Implement a newtype mechanism.
Unlike Haskell's `newtype` this doesn't introduce a constructor for the wrapper
(we don't have user-defined constructors generally) it just introduces a new
type that's isomorphic to an existing one but distinct from it. We can build
functions to map between the new type and the original one using an 'unsafe'
cast. We could introduce the mappings automatically, with syntax like
`newtype SpecialInt = Int toSpecialInt fromSpecialInt` but for now it's manual.