Improve error messages for class and linearity errors.
We do this by keeping source annotations around for the post-inference type
checking pass, where class and linearity errors are caught. We also defer
the `:t` command until after this pass.

The linearity errors are now ok but the class errors are still not very good,
because the error is detected at type application, which doesn't have any
obvious provenance in the source program. To fix this, we really need to move
class checking into the inference pass.