Fix another "atomic for" bug, similar to #30.
This one was noticed by @oxinabox in #33. The problem is that we need to expand
`NAtomicFor` into `NScan` before the `Imp` pass, but they've been slipping
through in a few places. 9bcc3e9f95 handles them in table constructors. This
commit handles them in arguments to `select` and the carry of `NScan`.
Generally, we need to take care whenever we have a `Data` constraint. We're
likely still missing some cases.

This change creates extra copies in some places (see examples/flop-tests.dx).
We'll worry about those later.