Rework simplification, especially decomposition.
The main goal was to handle tables of functions as outputs of `scan` but it
turned into a bigger refactor. The invariants are a bit clearer now. Previously
we created partially simplified expressions (`NExpr`), which we then decomposed
into data-only components and the 'ions' that can reconstruct functions. Now we
make sure to only create fully simplified expressions in the first place, with
all the non-data structure held in atoms.