Allow line breaks in more places and get rid of the `..` escaping mechanism.
We now treat a newline as whitespace whenever we can't interpret it as a decl
terminator. This includes within parens/brackets, or immediately after an infix
operator. Decls and parens can be arbitrarily nested so we need to maintain a
stack of "can we consume newlines" contexts. The reader monad handles this
nicely.

Fixes #45.