Add tuples to NExpr and share `Type` data type with Expr.
This is the first part of a multipart rewrite. The goal is to make it easier to
write new passes by having a single core IR, NExpr, that extends from just after
type inference to LLVM codegen. This commit extends Nexpr "up" by giving it
tuples/records and sharing the `Type` data type with the source-language `Expr`
IR. I hope to extend it "down" using monads to express increasingly stateful
computations.