Introduce a dedicated array type.
Previously we'd been treating arrays as a special case of tables. It was always
a source of confusion in the lowering to Imp/LLVM. For example, it forced us to
have conventions like `loadIfScalar`. And it was only going to get worse once we
added more complicated index sets.

Arrays and tables each have a single constructor, `ArrayRef` and `AFor`
respectively. `Afor` is the bridge between the array and table worlds. It
constructs a table out of arrays.