Bring back custom linearization rules.
As before, there's some trickiness in interleaving differentiation and
simplification. We also have to be careful not to inline functions with custom
derivatives too early. The situation might become clearer as we expand AD
coverage to the whole language and as we see some more examples of custom rules
(e.g. transposition).