Force linking of libdex.so by pretending to use one of its functions.
I can't figure out how else to get GHC to link it on MacOS (seems to be ok on
Linux). The functions in libdex.so are used in code generated by the JIT
compiler at runtime, but they don't otherwise appear in the Haskell source that
GHC knows about.

Fixes #56.