Switch to a Koka-like type-and-effects system instead of a monadic one.
Effects are now associated with functions rather than being standalone values.
Arrow types include an effect, which might be the empty effect:

    <type> ::= ... | <type> -> <effect> <type>

We can still mimic standalone effects with a nullary function. Effects are
carried out at function application. Currently the allowable effects are "no
effect" or a monolithic reader-writer-state effect, but we plan to break
this monolith into pieces and possibly add others.

=== Why ===

Effects are increasingly important, (1) in AD (for sparse accumulation), (2) as
a compilation target, and (3) for user-level in-place array algorithms. With the
monadic effect system, we had to implement everything twice, see e.g.
`toImpExpr` and `toImpActionExpr`. We had to handle ordinary pure programs, and
we also had to handle "monadic values", which were really just programs too.

Now, effects are part of ordinary programs, as they would be in an imperative
language. But they're still captured just as explicitly in the type system. So
I *think* this is essentially a syntactic change, in the sense that we can use
purely local rules to translate between the version of the language with monadic
effects and this new version.