Add some pattern synonyms for types and atoms.
The convention is that *Ty is the synonym and *Type is the true constructor
("the shorthand is shorter"):

    pattern TabTy a b = TC (TabType a b)

This reduces the burden of `Con`/`TC` wrapping and replaces it with the burden
of working with similarly-named constructors. I imagine it might make things
easier to read but harder to write. We'll see if it turns out to be a net win.