Bring back `:time` command and bring "benchmark.dx" up to date.
Timing things exposes a bug we already knew about: indexing subtables has been
done by copying ever since we made table-indexing an operation instead of an
atom. An easy fix would be to switch the Imp lowering from a caller-allocates
style to a callee-allocates style. That will be necessary anyway when we want to
materialize dependent/ragged tables.