Implement UExpr versions of `for`/`get`/`=>`.
I tried making these just primitive constants, just like `(+)`. It would look
something like this:

    def (=>) (n:Type) (a:Type) : Type = %TabType n a
    def (.)  {n:Type} {a:Type} (xs: n => a) (i:n) : a = %TabGet xs i
    def for  {n:Type} {a:Type} (f: n -> a) : n => a = %for f

The problem is that this doesn't let us have dependent table arrows. In the
above, we can't have an `a` that depends on the index of type `n`. Adam pointed
out that we could make it work if the type argument `a` was actually a
type-level function (of `i:n`), and maybe you could even infer the function. For
now, I'm leaving for/get as builtins, with very similar inference rules to
lambda/application.

We have four different function-like things now:

    a -> b   -- plain function
    a --o b  -- linear function
    a => b   -- tabulated function
    {a} -> b -- implicit function

In UExpr I'm trying to treat them more uniformly than we used to. Each of the
three function-related constructs (lambda, application and the arrow type
constructor) takes an `ArrowHead` parameter to specify which flavor of arrow
we're talking about.