Simplify type checker and make internal errors more informative.
We'd been trying to use one set of per-primitive typing rules to do three
things: checking, inference, and check-free querying, by abstracting over how to
compare types (check equality, generate constraints, or do nothing). With this
latest refactor we're no longer doing any type inference on primitives (we
assume they're trivially wrapped in the prelude). That still leaves checking and
querying, but they're already closer: we can just carry an optional typing
environment and skip the checks if it's absent. We're also able to unify type
checking and kind checking, which simplifies things further.

The better errors are because we're using do/monadfail to unpack type-promised
patterns (e.g. `Arrow a b <- ...` for the lhs of a function application). These
are softer than the hard-failing pattern matches we used to use so we can add
context before re-raising them.