Improve error messages.
Improve parse errors by adding a few more labels. Improve type errors by
renaming the inference variables like `?_3`, `?_15` to `a`, `b` etc., and
reporting the types of the variables being solved for. This should also make the
error text less sensitive to the internal details of type inference, so we don't
have to change the expected test outputs so often.