Define Ord instances in the prelude.
That Ord implies Eq is expressed by providing the projection for the Eq
component of the Ord dictionary:

    -- ("instance" isn't quite the right term. "synth hint" might be better)
    @instance
    def eqFromOrd (d:Ord a) ?=> : Eq a = fst (%fromClassDict d)

But there's a problem. If you define both `Ord Int` and `Eq Int` as instance
functions, then the synthesizer finds two ways to construct an `Eq Int`: the one
you defined directly, but also the one using `eqFromOrd`. The way I work around
this in this commit is to only define the `Ord` version as a hintable instance.
But that doesn't get us very far. For example, for pairs you'd like to define
both `Eq` and `Ord` so that pairs of equatable-but-not-ordered things can be
equated. I'm sure the solution is well known, but I haven't been able to figure
it out.