Add an imperative `while` primitive and an example Newton solver.
    while : (Unit -> {|eff} Bool) -> (Unit -> {|eff} Unit) -> {|eff} Unit

This will be handy for things like rejection samplers and iterative
optimizations, where the number of loop iterations isn't known ahead of time.
It's a bit of an odd duck among Dex looping constructs but the use cases seem
important enough to justify it. An alternative would be to just bite the bullet
and add recursion.

The loop is unproductive (it doesn't yield anything), so it's only used for its
effects. We'll need to make it productive in order to make it closed under
differentiation. That will require dependent pairs, since we only learn the size
of the resulting table after we're finished.