Add vector base types to support SIMD code generation (#127)
This is a first step that should allow us to play around with generating
reasonable single-threaded code for matrix multiplication. What was
previously called a `BaseType` is not a `ScalarBaseType`, while
`BaseType` is either a `Scalar` or `Vector` wrapper around that.
For now vectors always have a fixed width that is hardcoded as a
constant into the compiler as `vectorWidth` (in `Array.hs`).

There are two alternatives that I have considered instead of taking this
approach, but together with @dougalm we've decided that this seems to be
the best way forward for now. The options considered were:
1. Make vector types simply a newtype for `VectorWidth=>Real` that would
   get treated specially starting from Imp. That might work, but at the
   same time it seems a little misguided given that we really don't want
   those values to have the same representation, as vectors should
   always be stored in as atomic values in virtual registers, as opposed
   to arrays which will be usually maniupulated as references/pointers.
2. Don't introduce any new types for the vectors and make all SIMD
   instructions builtins that operate on array arguments. This would be
   nice in that we wouldn't have too many additions, but at the same
   time we _really_ want those fixed-width vectors to be represented
   differently, so they warrant having a different type.

I kind of like option 2., and I think that what we will eventually have
to do is add support for memory hierarchy such that we can e.g. ask to
have certain arrays allocated in the register file, instead of always
assuming that arrays live on the heap. This would be similar to
Fireiron, although our use case is considerably more complex given that
we might not always know the allocation size ahead of time, which is
required for register reservation.