Add an experimental LLVM-CUDA backend
This is the first of a series of patches that add a GPU backend for Dex.
For now, we assume that the outer loop in every program is parallel, and
distribute its iterations over GPU threads. Programs with nested loops
might not work too well just yet and might require flattening them into
a single loop over tuple index sets.

The backend itself does not work yet, but it is capable of converting a
sequential Imp program into a multi-device Imp IR where kernel launches
are explicit. It is also able to generate valid LLVM NVPTX IR and
compile it down to PTX assembly.