Allow ADTs with parametric types.
Pattern matching turned out to be trickier than expected! My first attempt
looked this:

    data Pair : Type -> Type -> Type where
      MkPair : a:Type -> b:Type -> a -> b -> Pair a b

    case (MkPair Int Int 1 2) of
      MkPair (a:Type) (b:Type) (x:a) (y:b) ->
        x + y   -- doesn't typecheck!

We want to be able to type `x + y` with `x:Int` and `y:Int` but instead we just
have `x:a` and `y:b`. Somehow we need to avoid binding those parameters (`a` and
`b`), whose values are already known from the type of the scrutinee.

A solution, suggested by Adam, is to allow the parameters to the type
constructor to be implicit parameters to the data constructors too. These
parameters are distinguished in that pattern constructors don't take them as
arguments. This works well, but it imposes the restriction that all data
constructors produce a type of the form `TC a1 a2 ..`, which is true of ordinary
ML/Haskell data constructors, but not GADTs or data types in Coq, Idris etc.
It seems like an acceptable restriction for now.