Complete the prototype of the LLVM-CUDA backend
This is the second half of the previous patch, which extends the LLVM
code generator to allow it to generate host code to dispatch CUDA
kernels. So far it's not been tested very thoroughly, so it's likely
that there are many rough edges and problems to be found, but I've
verified that it should be able to execute some simple pointwise maps.