Remove the builtin Int and Float types
After some discussion it became clear that this doesn't seem to be a
good approach and only adds unnecessary complexity. We keep the
distinguished preferred bit-width in the prelude for now, because our
implementation is not generic enough to be able to seamlessly work with
both types, but we should progressively be fixing this.