Pattern matching for extensible variants.
Matching on extensible variants was a bit complex, and I took a few
shortcuts. The main complication is that, given an extensible record,
it's not immediately clear how many alternatives there should be, since
it could be extended later (even during inference!). Thus, we want to
emit a VariantSplit to reduce it to the specific types we care about
and extract the polymorphic tail if needed. But the types we care about
can depend on what patterns we have!

The solution I implemented was to allow only one variant tail pattern
in a case expression, and require that it not overlap with any non-tail
variant patterns. This tail pattern is used to determine how to split
the scrutinee before we pattern match, and after splitting we then have
two disjoint non-extensible variants that we can pattern match on in
two steps.

Since this makes the canonical ordering of the alternatives a bit harder
to infer in advance, I've adapted the case pattern-match machinery to
use a helper ADT that tracks what kind of pattern we are matching, with
different cases for Dex ADT constructors, variant patterns, and variant
tail patterns.