Fix a subtle double-simplify bug.
I didn't realize this at first, but simplifyAtom/simplifyExpr are not
idempotent! In particular, simplification uses a different space of
variables (although they share the same namespace), so we might have
a subst env that binds X to Y and Y to Z. Simplifying twice can mean
that you bind X to Z by mistake!