Make the inliner use aggregation more accurate
The previous approach was quite eager to overestimate the number of uses
of each binder, because we tend to aggressively reuse names in
incomparable scopes. This patch adds an extra slot in each binder, which
can be filled in with inlining hints. The whole procedure is very
similar to DCE and the approach is heavily inspired by the "Secrets of
the GHC inliner" paper.