Improve defunctionalization of case statements
The previous method of extracting the closure necessary to rebuild the
atom after a case statement (find the set of free variables, emit an
acase with the same type as the case) was simple, but unfortunately it
caused the code we emitted later to get quite ugly. It made us emit
another case statment for each output of the case, even the
data-conforming ones.

With this patch, we traverse the product type returned from the case and
gather all the leaf values. All the data-conforming leaves are returned
as they are, since their value can be resolved within this single
branch set. All non-data leaves cause us to emit a "closure value"
indicating which case has been taken and capturing all the local values
necessary to reconstruct the non-data atom. This way we only emit extra
case statements at the use-site of non-data values, which generally are
a relatively small subset of the result.

One important property we get from this is that doing AD on data-only
case statements preserves the number of `case` expressions in the primal
function.