Add reference types for types with dependent constructors.
This starts to enable codegen for lists! We have two new constructors for
`Atom`: `DataConRef` and `BoxedRef`. `DataConRef` holds references to ADTs with
a single possibly-dependent constructor. Its references to constructor args can
refer to dereferenced arg references from earlier args. `BoxedRef` holds a
pointer, an allocation size, and a reference, and it makes the pointer's
referent (another pointer) available in the reference. We can use `DataConRef`
and `BoxedRef` to create references for types whose runtime sizes are not known
from the type alone. For example, a reference to a `Fin 10 => List Float` might
look something like this:

```
\for i:(Fin 10).
  DataConRef List Float
      n:Int             <- nPtr +> ordinal i
      xs:(Fin n=>Float) <- BoxedRef
                             xsPtr:(Ptr Float) {size=n} <- xsPtrPtr +> ordinal i
                             \for j:(Fin n). xsPtr +> ordinal j
```