Implement transposition for zero table atoms
This was a case that could have been triggered when differentiating a
function with some outputs that don't depend on any active inputs, or in
the `withState` example attached in `ad-tests.dx`. Fixing it also
exposed a bug caused by an overly eager inlining in Embed, which did not
preserve linearity.