Implement allocation lifting for CUDA kernels
GPUs don't support dynamic memory allocation from inside the kernels, so
any non-stack allocations should be lifted to the host. This is an
initial version of this transformation, which only works for allocations
of statically known size. Extending it to arbitrary (loop-independent)
arithmetic would require us to lift whole blocks out of the loops, which
is a form of loop invariant code motion (size calculation would have to
happen on the host), and that seems like a natural next step now.

Here's an example program which would previously fail to compile (due to
the large intermediate state allocated in each thread), but should work
just fine now:
```
allocationLiftingTest =
  for i:(Fin 100).
    snd $ withState (for j:(Fin 1000). ordinal i) $ \s.
      s!(0@_) := get s!(0@_) + 1
```

Now that we stop treating Imp IR as a temporary stop along the way to
LLVM, I've figured out that it makes sense to start splitting the
directory structure of our compiler. The tooling we'll need for each of
our IRs is similar (e.g. we now have an Embed monad for both Core and
Imp), and organizing it into submodules can make the code a bit clearer.
Some TODOs here are:
* Lift the Imp IR definition from Syntax;
* Make the Core -> Imp translation use the new Embed monad;
* Implement a `MonadIEmbed` typeclass, similarly to Core's `MonadEmbed`.