Allow patterns in pi types in the untyped syntax.
UPi atoms now can take a pattern instead of a single binder. If the pattern
is more complex than a single binder, that pattern is then bound while
converting the UPi into a Pi atom. Note that the Pi representation is the
same as it was; the returned type must still be reducible to an atom for
the conversion to succeed. However, with the new projection atoms, unpacking
of ADTs will still be reducible.

The parser implementation for "def"-style functions has been modified to allow
using patterns, which means it is now possible to bring values from an ADT
into scope in the type for a "def"-style function. For now, the parser does not
support patterns in explicit pi type expressions. This should be fairly
straightforward to add but might require some care regarding ambiguity of the
grammar (see #282).