Handle `for` expressions when lowering away exceptions.
We want to preserve opportunities for parallelism even when we have exceptions.
This lowering does that by discharding the exception at each index, producing a
table of `Maybe`s (which can be evaluted in parallel) which are then combined.

This interacts with state in a slightly surprising way. If the loop is stateful,
we still run all the iterations (updating the state as we go) even if early
iterations fail.

It's much easier to implement `catMaybes : (n=>Maybe a) -> Maybe (n=>a)` in the
prelude instead of using Embed directly. So I also added a utility for calling
predlue-defined functions via Embed.