Continue inlining after application
Turns out that the inlining pass had a pretty big gaping hole
previously: the inlineable tables originate not only from the `for`
loops, but also from applications with table lambdas! Previously we
completely ignored the second case, potentially failing to fully
reducing a table application, and blowing up the run-time complexity as
in #346.