Use "bundles" instead of cons lists in parallelization
We always know the length of the cons list statically and the units add
a lot of syntactic noise. To make matters worse, our current
optimizations don't deal with them very well, so removing some cons
lists actually improves performance on certain benchmarks.

This is not the right fix long-term, but it gets us somewhere.